feat: extract wizard forms & add wizard submission reset and redirect
