--- src/features/job/add-job/components/steps/benefits/index.tsx
@@ -1,80 +1,0 @@
-import { type SubmitHandler, useForm } from "react-hook-form";
-
-import { Stack } from "@mui/material";
-
-import Divider from "@features/ui/Divider";
-import { useAppDispatch, useAppSelector } from "@store/index";
-
-import { type Job, type PerkBenefit } from "../../../../types";
-import {
-  nextStep,
-  selectJob,
-  setBenefits,
-} from "../../../store/jobWizardSlice";
-import Pagination from "../../navigation/Pagination";
-import CustomSelect from "../ui/CustomSelect";
-import { isCustomSelectObject } from "../utils";
-import BenefitCard from "./BenefitCard";
-import { PERKS_BENEFITS } from "./data";
-
-interface FormInput {
-  perksBenefits: Job["perksBenefits"];
-}
-
-export default function Benefits() {
-  const { handleSubmit, control, onSubmit } = useBenefitsForm();
-
-  return (
-    <Stack
-      component="form"
-      onSubmit={handleSubmit(onSubmit)}
-      noValidate
-      sx={{ width: "100%" }}
-      gap={3}
-    >
-      <CustomSelect
-        id="perksBenefits"
-        name="perksBenefits"
-        control={control}
-        title="Perks and Benefits"
-        subtitle="Encourage more people to apply by sharing the attractive rewards and benefits you offer your employees"
-        requireErrorText="perks and benefits"
-        buttonText="Benefit"
-        items={PERKS_BENEFITS}
-        renderSelectedItem={(item, removeItem) => {
-          return (
-            isCustomSelectObject(item) && (
-              <BenefitCard
-                key={item.id}
-                val={item as PerkBenefit}
-                onClose={() => removeItem(item)}
-              />
-            )
-          );
-        }}
-      />
-      <Divider />
-      <Pagination />
-    </Stack>
-  );
-}
-
-function useBenefitsForm() {
-  const job = useAppSelector(selectJob);
-  const { handleSubmit, control } = useForm<FormInput>({
-    defaultValues: {
-      perksBenefits: job.perksBenefits,
-    },
-  });
-  const dispatch = useAppDispatch();
-  const onSubmit: SubmitHandler<FormInput> = (data) => {
-    dispatch(nextStep());
-    dispatch(setBenefits(data.perksBenefits));
-  };
-
-  return {
-    handleSubmit,
-    control,
-    onSubmit,
-  };
-}
