+++ src/features/job/add-job/components/steps/jobDescription/types.ts
@@ -1,7 +1,5 @@
-import { FormInput } from ".";
-
 export interface FormField {
-  name: keyof FormInput;
+  name: string;
   title: string;
   subtitle: string;
   placeholder: string;
