+++ src/features/job/add-job/data.ts
@@ -3,10 +3,10 @@ import BallotOutlinedIcon from "@mui/icons-material/BallotOutlined";
 import BusinessCenterOutlinedIcon from "@mui/icons-material/BusinessCenterOutlined";
 import CardGiftcardOutlinedIcon from "@mui/icons-material/CardGiftcardOutlined";
 
-import Benefits from "./components/steps/benefits";
+import Benefits from "./components/steps/Benefits";
+import JobInfo from "./components/steps/JobInfo";
+import RecruitmentStages from "./components/steps/RecruitmentStages";
 import JobDescription from "./components/steps/jobDescription";
-import JobInfo from "./components/steps/jobInfo";
-import RecruitmentStages from "./components/steps/recruitmentStages";
 import type { WizardStep } from "./types";
 
 export const WIZARD_STEPS: WizardStep[] = [
