+++ src/features/job/ui/forms/benefits/data.ts
@@ -1,21 +1,4 @@
-import type { Job } from "../../../../types";
-import CameraIcon from "./assets/Camera.svg";
-import CharityIcon from "./assets/Charity.svg";
-import HealthIcon from "./assets/Health.svg";
-import TeaIcon from "./assets/Tea.svg";
-import TeamIcon from "./assets/Team.svg";
-import TrainIcon from "./assets/Train.svg";
-import VacationIcon from "./assets/Vacation.svg";
-
-export const PERKS_ICONS = {
-  CameraIcon,
-  CharityIcon,
-  HealthIcon,
-  TeaIcon,
-  TeamIcon,
-  TrainIcon,
-  VacationIcon,
-};
+import type { Job } from "../../../types";
 
 export const PERKS_BENEFITS: Job["perksBenefits"] = [
   {
