+++ src/features/job/ui/forms/index.ts
@@ -1,0 +1,10 @@
+export { default as CategoriesForm } from "./CategoriesForm";
+export { default as EmploymentTypesForm } from "./EmploymentTypesForm";
+export { default as TitleForm } from "./TitleForm";
+export { default as SalaryForm } from "./salary/SalaryForm";
+export { default as RequiredSkillsForm } from "./RequiredSkillsForm";
+export { default as TextInputForm } from "./TextInputForm";
+export { default as BenefitsForm } from "./benefits/BenefitsForm";
+export { default as StagesForm } from "./stages/StagesForm";
+export { default as NumberOfStagesForm } from "./stages/NumberOfStagesForm";
+export { default as CapacityForm } from "./stages/CapacityForm";
