+++ src/features/job/ui/forms/salary/ValueBlock.tsx
